implement filtering search item outcomes using logic

+++ DataSourceGenerator/LocationSourceGenerator.cs
@@ -16,12 +16,12 @@ namespace Data;
 public static class {locationsFileName}{{
     public readonly struct Entry(
         string name,
-        int apLocationId,
+        long apLocationId,
         int? clientId,
         string clientNameId
     ){{
         public readonly string name = name;
-        public readonly int apLocationId = apLocationId;
+        public readonly long apLocationId = apLocationId;
         public readonly int? clientId = clientId;
         public readonly string clientNameId = clientNameId;
     }}
@@ -99,10 +99,10 @@ public static class {locationsFileName}{{
         }
     }
 
-    public struct GeneratorLocation(string name, int locationId, int? clientId, string clientNameId, string[][] accessRules)
+    public struct GeneratorLocation(string name, long locationId, int? clientId, string clientNameId, string[][] accessRules)
     {
         public string name = name;
-        public int locationId = locationId;
+        public long locationId = locationId;
         public int? clientId = clientId;
         public string clientNameId = clientNameId;
         public string[][] accessRules = accessRules;
@@ -130,7 +130,7 @@ public static class {locationsFileName}{{
     struct Section
     {
         public string name;
-        public int location_id;
+        public long location_id;
         public string region;
         public int? client_id;
         public string client_name_id;
